Restructure sub-directories to be consistent between submodules and to avoid possibly package path collisions.
