+++ domain/src/main/java/org/folio/spring/domain/controller/advice/MultipleInterfaceAdvice.java
@@ -1,9 +1,9 @@
-package org.folio.spring.controller.advice;
+package org.folio.spring.domain.controller.advice;
 
-import org.folio.spring.controller.exception.SchemaIOException;
-import org.folio.spring.controller.exception.SchemaNotFoundException;
-import org.folio.spring.model.response.ResponseErrors;
-import org.folio.spring.utility.ErrorUtility;
+import org.folio.spring.domain.controller.exception.SchemaIOException;
+import org.folio.spring.domain.controller.exception.SchemaNotFoundException;
+import org.folio.spring.web.model.response.ResponseErrors;
+import org.folio.spring.web.utility.ErrorUtility;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
