+++ domain/src/main/java/org/folio/spring/domain/controller/exception/SchemaNotFoundException.java
@@ -1,4 +1,4 @@
-package org.folio.spring.controller.exception;
+package org.folio.spring.domain.controller.exception;
 
 public class SchemaNotFoundException extends RuntimeException {
