+++ domain/src/main/java/org/folio/spring/domain/service/JsonSchemasService.java
@@ -1,19 +1,17 @@
-package org.folio.spring.service;
+package org.folio.spring.domain.service;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import org.folio.spring.controller.exception.SchemaNotFoundException;
+import org.folio.spring.domain.controller.exception.SchemaNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.ResourcePatternResolver;
 import org.springframework.stereotype.Service;
-
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 @Service
 public class JsonSchemasService {
