+++ domain/src/main/java/org/folio/spring/domain/service/RamlsService.java
@@ -1,4 +1,4 @@
-package org.folio.spring.service;
+package org.folio.spring.domain.service;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -6,9 +6,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
 import org.apache.commons.io.IOUtils;
-import org.folio.spring.controller.exception.SchemaNotFoundException;
+import org.folio.spring.domain.controller.exception.SchemaNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.support.ResourcePatternResolver;
