+++ web/src/main/java/org/folio/spring/web/resolver/TokenHeaderResolver.java
@@ -1,7 +1,7 @@
-package org.folio.spring.resolver;
+package org.folio.spring.web.resolver;
 
-import org.folio.spring.annotation.TokenHeader;
-import org.folio.spring.utility.AnnotationUtility;
+import org.folio.spring.web.annotation.TokenHeader;
+import org.folio.spring.web.utility.AnnotationUtility;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
